Admin and Client Side code without Web service code
Do not run this code since the MasterKey is removed for security
reasons. If you install the setup and run the installed software, it
will connect to webservice on Softultima.com and try to authenticate
you. Without which the software doesn't work. So just go ahead and look
at the code. if you need any more details, feel free to ask.